requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
jinja2>=3.1.2
httpx>=0.24.0 